requirements.txt
# Core Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0

# RFID Card Reading
pyscard>=2.3.1
nfcpy==1.0.4

# Encryption
cryptography==41.0.7

# Scheduling (Midnight Reset)
APScheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Logging & Utilities
colorlog==6.8.0

# Backup: Keyboard Automation (if needed)
pyautogui==0.9.54

# Testing
pytest==7.4.3
pytest-flask==1.3.0